Move the generic methods into the `ReviewData` class
